extracted request logging from the Operation
